refactor: Add the other settings to Settings.ts and refactor settings tab
